+++ cmd/thanos/query-frontend.go
@@ -16,10 +16,10 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/model"
 	"github.com/weaveworks/common/user"
 	"gopkg.in/alecthomas/kingpin.v2"
 
+	"github.com/cortexproject/cortex/pkg/util/validation"
 	"github.com/thanos-io/thanos/pkg/component"
 	"github.com/thanos-io/thanos/pkg/extflag"
 	"github.com/thanos-io/thanos/pkg/extprom"
@@ -27,82 +27,62 @@ import (
 	"github.com/thanos-io/thanos/pkg/logging"
 	"github.com/thanos-io/thanos/pkg/prober"
 	"github.com/thanos-io/thanos/pkg/queryfrontend"
-	"github.com/thanos-io/thanos/pkg/queryfrontend/cache"
 	httpserver "github.com/thanos-io/thanos/pkg/server/http"
 	"github.com/thanos-io/thanos/pkg/server/http/middleware"
 	"github.com/thanos-io/thanos/pkg/tracing"
 )
 
-type queryFrontendConfig struct {
-	http             httpConfig
-	queryRangeConfig queryRangeConfig
+type config struct {
+	http httpConfig
+	queryfrontend.Config
 
-	downstreamURL        string
-	compressResponses    bool
-	LogQueriesLongerThan time.Duration
-
+	contentOrURL           extflag.PathOrContent
 	requestLoggingDecision string
-}
-
-type queryRangeConfig struct {
-	respCacheConfig     extflag.PathOrContent
-	cacheMaxFreshness   time.Duration
-	splitInterval       model.Duration
-	maxRetries          int
-	maxQueryParallelism int
-	maxQueryLength      model.Duration
-
 	// partialResponseStrategy is the default strategy used
 	// when parsing thanos query request.
 	partialResponseStrategy bool
 }
 
-func (c *queryRangeConfig) registerFlag(cmd *kingpin.CmdClause) {
-	cmd.Flag("query-range.split-interval", "Split queries by an interval and execute in parallel, 0 disables it.").
-		Default("24h").SetValue(&c.splitInterval)
-
-	cmd.Flag("query-range.max-retries-per-request", "Maximum number of retries for a single request; beyond this, the downstream error is returned.").
-		Default("5").IntVar(&c.maxRetries)
-
-	cmd.Flag("query-range.max-query-length", "Limit the query time range (end - start time) in the query-frontend, 0 disables it.").
-		Default("0").SetValue(&c.maxQueryLength)
-
-	cmd.Flag("query-range.max-query-parallelism", "Maximum number of queries will be scheduled in parallel by the frontend.").
-		Default("14").IntVar(&c.maxQueryParallelism)
-
-	cmd.Flag("query-range.response-cache-max-freshness", "Most recent allowed cacheable result, to prevent caching very recent results that might still be in flux.").
-		Default("1m").DurationVar(&c.cacheMaxFreshness)
-
-	cmd.Flag("query-range.partial-response", "Enable partial response for queries if no partial_response param is specified. --no-query-range.partial-response for disabling.").
-		Default("true").BoolVar(&c.partialResponseStrategy)
-
-	c.respCacheConfig = *extflag.RegisterPathOrContent(cmd, "query-range.response-cache-config", "YAML file that contains response cache configuration.", false)
-}
-
-func (c *queryFrontendConfig) registerFlag(cmd *kingpin.CmdClause) {
-	c.queryRangeConfig.registerFlag(cmd)
-	c.http.registerFlag(cmd)
-
-	cmd.Flag("query-frontend.downstream-url", "URL of downstream Prometheus Query compatible API.").
-		Default("http://localhost:9090").StringVar(&c.downstreamURL)
-
-	cmd.Flag("query-frontend.compress-responses", "Compress HTTP responses.").
-		Default("false").BoolVar(&c.compressResponses)
-
-	cmd.Flag("query-frontend.log_queries_longer_than", "Log queries that are slower than the specified duration. "+
-		"Set to 0 to disable. Set to < 0 to enable on all queries.").Default("0").DurationVar(&c.LogQueriesLongerThan)
-
-	cmd.Flag("log.request.decision", "Request Logging for logging the start and end of requests. LogFinishCall is enabled by default. LogFinishCall : Logs the finish call of the requests. LogStartAndFinishCall : Logs the start and finish call of the requests. NoLogCall : Disable request logging.").Default("LogFinishCall").EnumVar(&c.requestLoggingDecision, "NoLogCall", "LogFinishCall", "LogStartAndFinishCall")
-}
-
 func registerQueryFrontend(m map[string]setupFunc, app *kingpin.Application) {
 	comp := component.QueryFrontend
 	cmd := app.Command(comp.String(), "query frontend")
-	conf := &queryFrontendConfig{}
-	conf.registerFlag(cmd)
+	cfg := &config{}
+
+	cmd.Flag("query-range.split-queries-by-interval", "Split queries by an interval and execute in parallel, 0 disables it.").
+		Default("24h").DurationVar(&cfg.QueryRange.SplitQueriesByInterval)
+
+	cmd.Flag("query-range.max-retries-per-request", "Maximum number of retries for a single request; beyond this, the downstream error is returned.").
+		Default("5").IntVar(&cfg.QueryRange.MaxRetries)
+
+	cmd.Flag("query-range.max-query-length", "Limit the query time range (end - start time) in the query-frontend, 0 disables it.").
+		Default("0").DurationVar(&cfg.Limits.MaxQueryLength)
+
+	cmd.Flag("query-range.max-query-parallelism", "Maximum number of queries will be scheduled in parallel by the Frontend.").
+		Default("14").IntVar(&cfg.Limits.MaxQueryParallelism)
+
+	cmd.Flag("query-range.max-cache-freshness", "Most recent allowed cacheable result, to prevent caching very recent results that might still be in flux.").
+		Default("1m").DurationVar(&cfg.Limits.MaxCacheFreshness)
+
+	cmd.Flag("query-range.partial-response", "Enable partial response for queries if no partial_response param is specified. --no-query-range.partial-response for disabling.").
+		Default("true").BoolVar(&cfg.partialResponseStrategy)
+
+	cfg.contentOrURL = *extflag.RegisterPathOrContent(cmd, "query-range.cache-config", "YAML file that contains response cache configuration.", false)
+
+	cfg.http.registerFlag(cmd)
+
+	cmd.Flag("query-frontend.downstream-url", "URL of downstream Prometheus Query compatible API.").
+		Default("http://localhost:9090").StringVar(&cfg.Frontend.DownstreamURL)
+
+	cmd.Flag("query-frontend.compress-http-responses", "Compress HTTP responses.").
+		Default("false").BoolVar(&cfg.Frontend.CompressResponses)
+
+	cmd.Flag("query-frontend.log-queries-longer-than", "Log queries that are slower than the specified duration. "+
+		"Set to 0 to disable. Set to < 0 to enable on all queries.").Default("0").DurationVar(&cfg.Frontend.LogQueriesLongerThan)
+
+	cmd.Flag("log.request.decision", "Request Logging for logging the start and end of requests. LogFinishCall is enabled by default. LogFinishCall : Logs the finish call of the requests. LogStartAndFinishCall : Logs the start and finish call of the requests. NoLogCall : Disable request logging.").Default("LogFinishCall").EnumVar(&cfg.requestLoggingDecision, "NoLogCall", "LogFinishCall", "LogStartAndFinishCall")
 
 	m[comp.String()] = func(g *run.Group, logger log.Logger, reg *prometheus.Registry, tracer opentracing.Tracer, _ <-chan struct{}, _ bool) error {
-		return runQueryFrontend(g, logger, reg, tracer, conf, comp)
+		return runQueryFrontend(g, logger, reg, tracer, cfg, comp)
 	}
 }
 
@@ -111,51 +91,40 @@ func runQueryFrontend(
 	logger log.Logger,
 	reg *prometheus.Registry,
 	tracer opentracing.Tracer,
-	conf *queryFrontendConfig,
+	cfg *config,
 	comp component.Component,
 ) error {
 
-	if len(conf.downstreamURL) == 0 {
-		return errors.New("downstream URL should be configured")
+	err := cfg.Cache.Validate()
+	if err != nil {
+		return errors.Wrap(err, "error validating cache config")
+	}
+	err = cfg.QueryRange.Validate(logger)
+	if err != nil {
+		return errors.Wrap(err, "error validating query range config")
 	}
 
 	fe, err := frontend.New(frontend.Config{
-		DownstreamURL:        conf.downstreamURL,
-		CompressResponses:    conf.compressResponses,
-		LogQueriesLongerThan: conf.LogQueriesLongerThan,
+		DownstreamURL:        cfg.Frontend.DownstreamURL,
+		CompressResponses:    cfg.Frontend.CompressResponses,
+		LogQueriesLongerThan: cfg.Frontend.LogQueriesLongerThan,
 	}, logger, reg)
 	if err != nil {
 		return errors.Wrap(err, "setup query frontend")
 	}
 	defer fe.Close()
 
-	limits := queryfrontend.NewLimits(
-		conf.queryRangeConfig.maxQueryParallelism,
-		time.Duration(conf.queryRangeConfig.maxQueryLength),
-		conf.queryRangeConfig.cacheMaxFreshness,
-	)
-
-	respCacheContentYaml, err := conf.queryRangeConfig.respCacheConfig.Content()
+	limits, err := validation.NewOverrides(cfg.Limits, nil)
 	if err != nil {
-		return errors.Wrap(err, "get content of response cache configuration")
+		return errors.Wrap(err, "initialiase limits")
 	}
 
-	var cacheConfig *queryrange.ResultsCacheConfig
-	if len(respCacheContentYaml) > 0 {
-		cacheConfig, err = cache.NewResponseCacheConfig(respCacheContentYaml)
-		if err != nil {
-			return errors.Wrap(err, "create response cache")
-		}
-	}
-
-	codec := queryfrontend.NewThanosCodec(conf.queryRangeConfig.partialResponseStrategy)
+	codec := queryfrontend.NewThanosCodec(cfg.partialResponseStrategy)
 	tripperWare, err := queryfrontend.NewTripperWare(
+		cfg.QueryRange,
 		limits,
-		cacheConfig,
 		codec,
 		queryrange.PrometheusResponseExtractor{},
-		time.Duration(conf.queryRangeConfig.splitInterval),
-		conf.queryRangeConfig.maxRetries,
 		reg,
 		logger,
 	)
@@ -173,7 +142,7 @@ func runQueryFrontend(
 
 	// Configure Request Logging for HTTP calls.
 	opts := []logging.Option{logging.WithDecider(func() logging.Decision {
-		return logging.LogDecision[conf.requestLoggingDecision]
+		return logging.LogDecision[cfg.requestLoggingDecision]
 	})}
 	logMiddleware := logging.NewHTTPServerMiddleware(logger, opts...)
 	ins := extpromhttp.NewInstrumentationMiddleware(reg)
@@ -181,8 +150,8 @@ func runQueryFrontend(
 	// Start metrics HTTP server.
 	{
 		srv := httpserver.New(logger, reg, comp, httpProbe,
-			httpserver.WithListen(conf.http.bindAddress),
-			httpserver.WithGracePeriod(time.Duration(conf.http.gracePeriod)),
+			httpserver.WithListen(cfg.http.bindAddress),
+			httpserver.WithGracePeriod(time.Duration(cfg.http.gracePeriod)),
 		)
 
 		instr := func(f http.HandlerFunc) http.HandlerFunc {