+++ cmd/thanos/compact.go
@@ -297,6 +297,7 @@ func runCompact(
 	}
 
 	var cleanMtx sync.Mutex
+	// TODO(GiedriusS): we could also apply retention policies here but the logic would be a bit more complex.
 	cleanPartialMarked := func() error {
 		cleanMtx.Lock()
 		defer cleanMtx.Unlock()
@@ -306,7 +307,22 @@ func runCompact(
 		if err := blocksCleaner.DeleteMarkedBlocks(ctx); err != nil {
 			return errors.Wrap(err, "error cleaning marked blocks")
 		}
+
+		if err := sy.SyncMetas(ctx); err != nil {
+			level.Error(logger).Log("msg", "failed to sync metas", "err", err)
+		}
 		return nil
+	}
+
+	// Do it once at the beginning to ensure that it runs at least once before
+	// the main loop.
+	if err := sy.SyncMetas(ctx); err != nil {
+		cancel()
+		return errors.Wrap(err, "syncing metas")
+	}
+	if err := cleanPartialMarked(); err != nil {
+		cancel()
+		return errors.Wrap(err, "cleaning partial and marked blocks")
 	}
 
 	compactMainFn := func() error {
@@ -427,13 +443,20 @@ func runCompact(
 
 		// Periodically remove partial blocks and blocks marked for deletion
 		// since one iteration potentially could take a long time.
-		g.Add(func() error {
-			return runutil.Repeat(5*time.Minute, ctx.Done(), func() error {
-				return cleanPartialMarked()
+		if conf.cleanupBlocksInterval > 0 {
+			g.Add(func() error {
+				// Wait the whole period at the beginning because we've executed this on boot.
+				select {
+				case <-time.After(conf.cleanupBlocksInterval):
+				case <-ctx.Done():
+				}
+				return runutil.Repeat(conf.cleanupBlocksInterval, ctx.Done(), func() error {
+					return cleanPartialMarked()
+				})
+			}, func(error) {
+				cancel()
 			})
-		}, func(error) {
-			cancel()
-		})
+		}
 
 		g.Add(func() error {
 			iterCtx, iterCancel := context.WithTimeout(ctx, conf.waitInterval)
@@ -477,6 +500,7 @@ type compactConfig struct {
 	disableDownsampling                            bool
 	blockSyncConcurrency                           int
 	blockViewerSyncBlockInterval                   time.Duration
+	cleanupBlocksInterval                          time.Duration
 	compactionConcurrency                          int
 	deleteDelay                                    model.Duration
 	dedupReplicaLabels                             []string
@@ -526,6 +550,8 @@ func (cc *compactConfig) registerFlag(cmd extkingpin.FlagClause) {
 		Default("20").IntVar(&cc.blockSyncConcurrency)
 	cmd.Flag("block-viewer.global.sync-block-interval", "Repeat interval for syncing the blocks between local and remote view for /global Block Viewer UI.").
 		Default("1m").DurationVar(&cc.blockViewerSyncBlockInterval)
+	cmd.Flag("compact.cleanup-interval", "How often we should clean up partially uploaded blocks and blocks with deletion mark in the background when --wait has been enabled. Setting it to \"0s\" disables it - the cleaning will only happen at the end of an iteration.").
+		Default("5m").DurationVar(&cc.cleanupBlocksInterval)
 
 	cmd.Flag("compact.concurrency", "Number of goroutines to use when compacting groups.").
 		Default("1").IntVar(&cc.compactionConcurrency)